services/commentservice/requirements.txt
Flask==1.1.4
Flask-GraphQL==2.0.1
graphene==2.1.9
graphql-core==2.3.2
graphene-sqlalchemy==2.2.0
mysql-connector-python
python-dotenv
